backend-fastapi/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic==0.43.1
python-dotenv==1.0.1
pydantic==2.10.5
